Updated the README.md. Updated the installer.bat file (also removed the option to use pyinstaller). Added instructions to the game to help the player. Added an icon to the .exe.

+++ src/game_mechanics/tap_faucet.py
@@ -10,7 +10,7 @@ from src.textures.model_texture import ModelTexture
 
 
 class TapFaucet:
-    def __init__(self, obj_loader, loader, pos: list[float], rotation: list[float], size: int, fill_texture, content: str,
+    def __init__(self, obj_loader, loader, pos: list[float], rotation: list[float], size: float, fill_texture, content: str,
                  sign_texture: str):
         self.__obj_loader = obj_loader
         self.__loader = loader
